requirements.txt
nameko-sqlalchemy
nameko
pymongo
sqlalchemy
mysql-python
pymongo
multiprocessing
mongolog
docopt
sphinx
pytest
tox
redis
mogo
PyYAML
fudge
mockredispy
mock
mocksqlalchemy
